RF: Slightly restructure tournament hierarchy.
